training/requirements.txt
gensim==4.1.2
pyyaml==6.0.2
